moved study set and user services
