+++ src/app/app.config.ts
@@ -3,21 +3,21 @@ import { provideRouter } from '@angular/router';
 import { routes } from './app.routes';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
 import { HttpClientModule } from '@angular/common/http';
-import { StudySetDevService } from './services/study-set-dev.service';
+import { StudySetDevService } from './services/study-set/study-set-dev.service';
 import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
 import { getAuth, provideAuth } from '@angular/fire/auth';
 import { connectFirestoreEmulator, getFirestore, provideFirestore } from '@angular/fire/firestore';
 import { connectStorageEmulator, getStorage, provideStorage } from '@angular/fire/storage';
 import { environment } from '../environments/environment';
-import { StudySetService } from './services/study-set.service';
+import { StudySetService } from './services/study-set/study-set.service';
 import { MATERIAL_SANITY_CHECKS } from '@angular/material/core';
-import { UserInfoService } from './services/user-info.service';
-import { UserInfoFakeService } from './services/user-info-fake.service';
+import { UserInfoService } from './services/user/user-info.service';
+import { UserInfoFakeService } from './services/user/user-info-fake.service';
 import { AuthService } from './services/auth/auth.service';
 import { AuthDevService } from './services/auth/auth-dev.service';
 import { AuthFirebaseService } from './services/auth/auth-firebase.service';
-import { StudySetFirebaseService } from './services/study-set-firebase.service';
-import { UserInfoFirebaseService } from './services/user-info-firebase.service';
+import { StudySetFirebaseService } from './services/study-set/study-set-firebase.service';
+import { UserInfoFirebaseService } from './services/user/user-info-firebase.service';
 import {
   SearchStudySetService,
   SearchStudySetTypesenseService,
