+++ src/app/services/user/user-info-fake.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
-import { AccessStorageData, UserData } from '../data-models/user-model';
+import { AccessStorageData, UserData } from '../../data-models/user-model';
 import { of, Observable } from 'rxjs';
 import { UserInfoService } from './user-info.service';
-import { StudySetData } from '../data-models/studyset-model';
+import { StudySetData } from '../../data-models/studyset-model';
 import { take } from 'rxjs';
